removing local selected views from canvas-component-entry, moving it back to new-canvas-controls
